requirements.txt
# Core dependencies - pinned versions for reproducibility
numpy==2.2.6
scipy==1.15.3
matplotlib==3.10.8
seaborn==0.13.2
pandas==2.3.3

# Deep Learning
torch==2.9.1
snntorch==0.9.4

# Experiment management
pyyaml==6.0.3
pydantic==2.12.5
pydantic-core==2.41.5
h5py==3.15.1
tqdm==4.68.0

# Visualization
plotly==6.5.1
networkx==3.4.2

# Testing
pytest==9.0.2
pytest-cov==7.0.0

# Notebooks
jupyter==1.1.1
jupyter-client==8.8.0
jupyter-core==5.9.1
jupyterlab==4.5.1
ipywidgets==8.1.8

# Utils
python-dateutil==2.9.0.post0

